spanner_dbapi: only keep autocommit_on for now
Moved the code with autocommit_off aka Transaction long-holding
code into a separate branch

    DO-NOT-DELETE-special-branch-with-autocommit-off

This way we can work with one coherent style and not
a partially fleshed implementation that's complex.

Fixes #381
